eventページのページ設計
