+++ client/pages/event/[id].tsx
@@ -2,22 +2,19 @@ import Head from 'next/head';
 import { Inter } from '@next/font/google';
 import { useState } from 'react';
 import { useRouter } from 'next/router';
+import { MyHead } from '../../components/templates/shared/MyHead';
+import { BasicTemplate } from '../../components/templates/shared/BasicTemplate';
 
 const inter = Inter({ subsets: ['latin'] });
 
-export default function Home() {
+export default function Show() {
   const [isHost, setIsHost] = useState(false);
   const [hasJoined, setHasJoined] = useState(false);
   const event_id = useRouter().query.id;
   return (
     <>
-      <Head>
-        <title>参加フォーム画面</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <main>
+      <MyHead title="募集タイトルを入れる" description="" />
+      <BasicTemplate className="text-center">
         <h1>参加フォーム画面</h1>
         <div>event id: {event_id}</div>
         <button onClick={() => setIsHost((v) => !v)}>
@@ -25,9 +22,7 @@ export default function Home() {
         </button>
 
         {isHost ? (
-          <>
-            <button>募集を締め切る</button>
-          </>
+          <button>募集を締め切る</button>
         ) : !hasJoined ? (
           <>
             <h2>名前</h2>
@@ -42,7 +37,7 @@ export default function Home() {
             <button onClick={() => setHasJoined(false)}> キャンセル </button>
           </>
         )}
-      </main>
+      </BasicTemplate>
     </>
   );
 }
